add checks for services

+++ check-network.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-if [ $# != 1 ]; then
-        echo "Usage: $0 port ## checks port for fullmesh connectivity in k8s cluster"
+if [ $# != 2 ]; then
+        echo "Usage: $0 port check-hostnetwork(yes|no)  ## checks port for fullmesh connectivity in k8s cluster"
         exit
 fi
 
@@ -37,6 +37,8 @@ sleep 10
 kubectl get pod -o wide | grep echoserver | awk '{print $1}' | xargs -I % -n 1 kubectl expose pod %  --port $1
 
 
+if [ "$2" == "yes" ] ; 
+then
 
 cat <<EOF2 | kubectl create -f -
 apiVersion: extensions/v1beta1
@@ -67,6 +69,8 @@ spec:
 EOF2
 
 
+fi
+
 
 
 
@@ -75,39 +79,42 @@ sleep 120
 
 export failures=''
 
-hosts=$(kubectl get pod -o wide | grep echoserver | awk '{print $6}')
-kubehosts=$(kubectl get pod -o wide | grep echoserver | awk '{print $1}')
+ips=$(kubectl get pod -o wide | grep echoserver | awk '{print $6}')
+pods=$(kubectl get pod -o wide | grep echoserver | awk '{print $1}')
 svcs=$(kubectl get svc -o wide | grep echoserver  | awk '{print $3}')
 
-for host in $hosts
+for pod in $pods
+
 do
 
 
- for kubehost in $kubehosts
+ for ip in $ips
  do
- echo "checking pod - $kubehost connection to host ${host}:${1}"
- kubectl exec $kubehost -- timeout 2 echo dummy-payload  > /dev/tcp/$host/$1 || failures="$failures \n pod - $kubehost cant connect to host $host:${1}"
+ echo "checking pod - $pod connection to host ${ip}:${1}"
+ kubectl exec $pod -- echo dummy-payload | nc -i 2 -4 -n --send-only $ip $1 || failures="$failures \n pod - $pod cant connect to host at $ip:${1}"
+ done
 
 
-  for svc in $svcs  
-  do
-  echo "checking pod - $kubehost connection to svc ${svc}:${1}" 
-  kubectl exec $kubehost -- timeout 2 echo dummy-payload  > /dev/tcp/$svc/$1 || failures="$failures \n pod - $kubehost cant connect to svc at $svc:${1}"
-    
-
-  done
-
+  
+ for svc in $svcs
+ do
+ echo "checking pod - $pod connection to svc ${svc}:${1}" 
+ kubectl exec $pod -- echo dummy-payload | nc -i 2 -4 -n --send-only $svc $1 || failures="$failures \n pod - $pod cant connect to svc at $svc:${1}"
  done
 
 done
 
 
 
+
 echo -e "failed hosts: $failures"
 
 
-
+if [ "$2" == "yes" ] ;
+then
 kubectl delete ds echoserver-hostnetwork
+fi
+
 kubectl delete ds echoserver
 kubectl get svc -o wide |  grep echoserver | awk '{print $1}' | xargs kubectl delete svc
